txn: split out process module
