+++ src/storage/txn/mod.rs
@@ -12,13 +12,15 @@
 // limitations under the License.
 
 mod latch;
+mod process;
 mod scheduler;
 mod store;
 
 use std::error;
 use std::io::Error as IoError;
 
-pub use self::scheduler::{Msg, Scheduler, CMD_BATCH_SIZE, RESOLVE_LOCK_BATCH_SIZE};
+pub use self::process::RESOLVE_LOCK_BATCH_SIZE;
+pub use self::scheduler::{Msg, Scheduler, CMD_BATCH_SIZE};
 pub use self::store::{SnapshotStore, StoreScanner};
 
 quick_error! {
